Typed annotated mixed list

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -6,7 +6,7 @@ Type annotated function that takes in a list of mixed type.
 from typing import Union
 
 
-def sum_mixed_list(mxd_lst: Union[float, int]) -> float:
+def sum_mixed_list(mxd_lst: Union[int, float]) -> float:
     """Returns sum of a mixed type of list as float"""
     total: float = 0.00
     for elem in mxd_lst:
